// src/types/WooCommerceTypes.ts
export interface Attribute_Variation {
  id: number;
  name: string;
  option: string;
}